requirements.txt
fastapi
uvicorn
openai
pinecone-client
pdfplumber
langchain
python-dotenv
tiktoken